+++ src/main/java/tn/esprit/projetpiback/entites/Evenement.java
@@ -17,7 +17,7 @@ import java.util.List;
 public class Evenement {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idEvenement;
+    private Long idEvenement;
     private Date dateDebut;
     private Date dateFin;
     private int nbrMaxParticipants;
