+++ src/main/java/tn/esprit/projetpiback/services/EvenementService.java
@@ -3,7 +3,7 @@ package tn.esprit.projetpiback.services;
 import tn.esprit.projetpiback.entites.Reservation;
 
 public interface EvenementService {
-    void insertEvenementWithNbrParticipants(Integer idEvenement);
+    void insertEvenementWithNbrParticipants(Long idEvenement);
 
-    Reservation addReservationAndAssignToEvenement(Integer idUser, Integer idEvenement);
+    Reservation addReservationAndAssignToEvenement(Integer idUser, Long idEvenement);
 }
